Move `MockClock` to `janus_core::time::test_utils`
